refactor(OpenApi3.1.x-Yaml): base adapter on generic YAML adapter
Refs https://github.com/swagger-api/oss-planning/issues/133
