+++ apidom/packages/apidom-parser-adapter-asyncapi-json-2-0/package.json
@@ -28,6 +28,7 @@
     "apidom": "file:../apidom",
     "apidom-ast": "file:../apidom-ast",
     "apidom-ns-asyncapi-2-0": "file:../apidom-ns-asyncapi-2-0",
+    "apidom-parser-adapter-json": "file:../apidom-parser-adapter-json",
     "ramda": "=0.27.0",
     "@types/ramda": "=0.27.6",
     "ramda-adjunct": "=2.27.0",
