+++ apidom/packages/apidom-parser-adapter-openapi-json-3-1/src/parser/index.ts
@@ -1,6 +1,13 @@
 import $RefParser from '@apidevtools/json-schema-ref-parser';
 import { createNamespace, ParseResultElement } from 'apidom';
-import { transformTreeSitterJsonCST } from 'apidom-ast';
+import {
+  Error,
+  JsonArray,
+  JsonDocument,
+  JsonObject,
+  JsonProperty,
+  transformTreeSitterJsonCST,
+} from 'apidom-ast';
 import openapi3_1 from 'apidom-ns-openapi-3-1';
 // @ts-ignore
 import { visit } from 'apidom-parser-adapter-json';
@@ -23,7 +30,21 @@ const parse = async (
   const cst = parser.parse(source);
   const ast = transformTreeSitterJsonCST(cst);
 
+  const keyMap = {
+    // @ts-ignore
+    [JsonDocument.type]: ['children'],
+    // @ts-ignore
+    [JsonObject.type]: ['children'],
+    // @ts-ignore
+    [JsonProperty.type]: ['children'],
+    // @ts-ignore
+    [JsonArray.type]: ['children'],
+    // @ts-ignore
+    [Error.type]: ['children'],
+  };
+
   visit(ast.rootNode, documentVisitor, {
+    keyMap,
     // @ts-ignore
     state: {
       namespace,
