--- apidom/packages/apidom-parser-adapter-openapi-yaml-3-1/src/parser/metadata.ts
@@ -1,11 +1,0 @@
-import { Element } from 'minim';
-
-// eslint-disable-next-line import/prefer-default-export
-export const appendMetadata = <T extends Element>(metadata: string[], element: T): T => {
-  metadata.forEach((md: string) => {
-    element.classes.push(md);
-    element.getMetaProperty('symbols', []).push(md);
-  });
-
-  return element;
-};
