+++ apidom/packages/apidom-parser-adapter-openapi-yaml-3-1/src/parser/visitors/DocumentVisitor.ts
@@ -1,34 +1,13 @@
 import stampit from 'stampit';
-import { YamlComment, YamlDocument, YamlMapping, YamlScalar, YamlSequence } from 'apidom-ast';
+import { YamlMapping } from 'apidom-ast';
+// @ts-ignore
+import { DocumentVisitor as YamlDocumentVisitor } from 'apidom-parser-adapter-yaml-1-2';
 
-import SpecificationVisitor from './SpecificationVisitor';
-
-const DocumentVisitor = stampit(SpecificationVisitor, {
-  props: {
-    keyMap: {
-      // @ts-ignore
-      [YamlDocument.type]: ['children'],
-    },
-  },
+const DocumentVisitor = stampit(YamlDocumentVisitor, {
   methods: {
-    scalar(scalarNode: YamlScalar) {
-      const element = this.nodeToElement(['scalar'], scalarNode);
-      this.element.content.push(element);
-    },
-
     mapping(mappingNode: YamlMapping) {
       const openApiElement = this.nodeToElement(['document', 'objects', 'OpenApi'], mappingNode);
       this.element.content.push(openApiElement);
-    },
-
-    sequence(sequenceNode: YamlSequence) {
-      const arrayElement = this.nodeToElement(['sequence'], sequenceNode);
-      this.element.content.push(arrayElement);
-    },
-
-    comment(commentNode: YamlComment) {
-      const commentElement = new this.namespace.elements.Comment(commentNode.content);
-      this.element.content.push(commentElement);
     },
   },
 });
