+++ apidom/packages/apidom-parser-adapter-openapi-yaml-3-1/src/parser/visitors/generics/MapYamlMappingVisitor.ts
@@ -1,12 +1,2 @@
-import stampit from 'stampit';
-import { isNonEmptyString } from 'ramda-adjunct';
-
-import PatternedFieldsYamlMappingVisitor from './PatternedFieldsYamlMappingVisitor';
-
-const MapYamlMappingVisitor = stampit(PatternedFieldsYamlMappingVisitor, {
-  props: {
-    fieldPatternPredicate: isNonEmptyString,
-  },
-});
-
-export default MapYamlMappingVisitor;
+// @ts-ignore
+export { MapYamlMappingVisitor as default } from 'apidom-parser-adapter-yaml-1-2';
