+++ apidom/packages/apidom-parser-adapter-openapi-yaml-3-1/src/parser/visitors/generics/MixedFieldsYamlMappingVisitor.ts
@@ -1,42 +1,12 @@
 import stampit from 'stampit';
-import { noop } from 'ramda-adjunct';
-import { YamlMapping } from 'apidom-ast';
+// @ts-ignore
+import { MixedFieldsYamlMappingVisitor as BaseVisitor } from 'apidom-parser-adapter-yaml-1-2';
 
-import { BREAK, visit } from '..';
-import FixedFieldsYamlMappingVisitor from './FixedFieldsYamlMappingVisitor';
-import PatternedFieldsYamlMappingVisitor from './PatternedFieldsYamlMappingVisitor';
-import SpecificationVisitor from '../SpecificationVisitor';
+import { isOpenApiExtension } from '../../predicates';
 
-const MixedFieldsYamlMappingVisitor = stampit(SpecificationVisitor, {
+const MixedFieldsYamlMappingVisitor = stampit(BaseVisitor, {
   props: {
-    specPathFixedFields: noop,
-    specPathPatternedFields: noop,
-  },
-  methods: {
-    mapping(mappingNode: YamlMapping) {
-      const fixedFieldsVisitor = FixedFieldsYamlMappingVisitor({
-        ...this.retrievePassingOptions(),
-        ignoredFields: this.ignoredFields,
-        canSupportSpecificationExtensions: this.canSupportSpecificationExtensions,
-        element: this.element,
-        specPath: this.specPathFixedFields,
-      });
-
-      visit(mappingNode, fixedFieldsVisitor);
-
-      const patternedFieldsVisitor = PatternedFieldsYamlMappingVisitor({
-        ...this.retrievePassingOptions(),
-        ignoredFields: this.ignoredFields,
-        canSupportSpecificationExtensions: this.canSupportSpecificationExtensions,
-        element: this.element,
-        fieldPatternPredicate: this.fieldPatternPredicate,
-        specPath: this.specPathPatternedFields,
-      });
-
-      visit(mappingNode, patternedFieldsVisitor);
-
-      return BREAK;
-    },
+    specificationExtensionPredicate: isOpenApiExtension({}),
   },
 });
