+++ apidom/packages/apidom-parser-adapter-openapi-yaml-3-1/src/parser/visitors/open-api-3-1/ServersVisitor.ts
@@ -1,8 +1,8 @@
 import stampit from 'stampit';
 import { YamlSequence } from 'apidom-ast';
+// @ts-ignore
+import { BREAK, SpecificationVisitor } from 'apidom-parser-adapter-yaml-1-2';
 
-import { BREAK } from '..';
-import SpecificationVisitor from '../SpecificationVisitor';
 import { isServerObject } from '../../predicates';
 import { KindVisitor } from '../generics';
