+++ apidom/packages/apidom-parser-adapter-openapi-yaml-3-1/src/parser/visitors/open-api-3-1/info/VersionVisitor.ts
@@ -1,8 +1,9 @@
 import stampit from 'stampit';
 import { YamlScalar } from 'apidom-ast';
+// @ts-ignore
+import { appendMetadata } from 'apidom-parser-adapter-yaml-1-2';
 
 import { KindVisitor } from '../../generics';
-import { appendMetadata } from '../../../metadata';
 
 const VersionVisitor = stampit(KindVisitor, {
   methods: {
