+++ apidom/packages/apidom-parser-adapter-openapi-yaml-3-1/src/parser/visitors/open-api-3-1/path-item/$RefVisitor.ts
@@ -1,8 +1,8 @@
 import stampit from 'stampit';
 import { YamlScalar } from 'apidom-ast';
+// @ts-ignore
+import { BREAK, SpecificationVisitor } from 'apidom-parser-adapter-yaml-1-2';
 
-import SpecificationVisitor from '../../SpecificationVisitor';
-import { BREAK } from '../..';
 import { KindVisitor } from '../../generics';
 
 const $RefVisitor = stampit(KindVisitor, SpecificationVisitor, {
